feat: 플레이리스트 단건 조회 구현

+++ src/main/java/org/example/mopl/content/dto/response/ContentDto.java
@@ -18,7 +18,7 @@ public record ContentDto(
     String thumbnailUrl,
     List<String> tags,
     Double averageRating,
-    Integer reviewCount,
+    Long reviewCount,
     Long watcherCount
 ) {
 
@@ -31,7 +31,7 @@ public record ContentDto(
         String thumbnailUrl,
         List<String> tags,
         Double averageRating,
-        Integer reviewCount,
+        Long reviewCount,
         Long watcherCount
     ) {
         return ContentDto.builder()
